Introduce new module Unit

+++ src/Modules/Core/DTO/GeoCoordinate.php
@@ -16,7 +16,7 @@ class GeoCoordinate
 	/**
 	 * Longitude.
 	 */
-	public int $lon = 0;
+	public float $lon = 0;
 
 	/**
 	 * Generates from an array like a DB result
